refactor: move tests to tools_test.go
